backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
python-jose[cryptography]==3.3.0
requests==2.31.0
pyyaml==6.0.1
minio==7.2.0
django-filter==23.5
drf-spectacular==0.27.0
weasyprint==61.2
pydyf==0.10.0
